Split TimeSignature and NoteValue files
